Update README.md
This project focuses on utilizing unsupervised machine learning techniques to predict the optimum number of clusters in the 'Iris' dataset. The 'Iris' dataset is a well-known dataset in the field of machine learning, consisting of measurements of different features of Iris flowers, including sepal length, sepal width, petal length, and petal width. By applying the elbow method and k-means clustering algorithm, we aim to determine the ideal number of clusters that best represent the underlying patterns and variations in the dataset. Scaling the data using the StandardScaler ensures that all features are on a similar scale, allowing for more meaningful comparisons. The elbow method aids in identifying the optimal number of clusters by evaluating the within-cluster sum of squares. Based on this analysis, the k-means clustering model is trained using the determined number of clusters. The resulting clusters are then visualized using a scatter plot, allowing us to observe the separation of data points among different clusters and gain insights into the inherent structure of the data. Additionally, an optional step involves comparing the predicted clusters with the actual target labels, if available, to assess the performance of the clustering algorithm. By exploring this project, we can gain a deeper understanding of unsupervised learning techniques and their application in identifying clusters within a dataset, providing a foundation for further analysis and pattern recognition in various real-world scenarios.